+++ Plugin.Notifications.tvOS/Properties/AssemblyInfo.cs
@@ -1,4 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("Plugin.Notifications.tvOS")]
-[assembly: AssemblyDescription("")]